minor spec modif

+++ specs/SyncSpec/SynchronizerSpec.cpp
@@ -35,21 +35,21 @@ go_bandit([](){
 
 			stripClock.setTime(9600);
 
-			AssertThat((int)stripClock.time(), Equals(9600));
-			AssertThat((int)videoClock.time(), Equals(9600));
-			AssertThat((int)syncClock.time(), Equals(0));
+			AssertThat(stripClock.time(), Equals(9600));
+			AssertThat(videoClock.time(), Equals(9600));
+			AssertThat(syncClock.time(), Equals(0));
 
 			sync.setSyncClock(&syncClock, PhSynchronizer::Sony);
 
 			stripClock.setTime(960);
 
-			AssertThat((int)stripClock.time(), Equals(960));
-			AssertThat((int)videoClock.time(), Equals(9600));
-			AssertThat((int)syncClock.time(), Equals(0));
+			AssertThat(stripClock.time(), Equals(960));
+			AssertThat(videoClock.time(), Equals(9600));
+			AssertThat(syncClock.time(), Equals(0));
 			stripClock.setTime(1920);
-			AssertThat((int)stripClock.time(), Equals(0));
-			AssertThat((int)videoClock.time(), Equals(9600));
-			AssertThat((int)syncClock.time(), Equals(0));
+			AssertThat(stripClock.time(), Equals(0));
+			AssertThat(videoClock.time(), Equals(9600));
+			AssertThat(syncClock.time(), Equals(0));
 		});
 
 		it("handle_strip_rat_e_change", []() {
@@ -84,9 +84,9 @@ go_bandit([](){
 
 			videoClock.setTime(9600);
 
-			AssertThat((int)stripClock.time(), Equals(0));
-			AssertThat((int)videoClock.time(), Equals(9600));
-			AssertThat((int)syncClock.time(), Equals(0));
+			AssertThat(stripClock.time(), Equals(0));
+			AssertThat(videoClock.time(), Equals(9600));
+			AssertThat(syncClock.time(), Equals(0));
 		});
 
 		it("handle_video_rate_change", []() {
@@ -114,22 +114,22 @@ go_bandit([](){
 
 			syncClock.setTime(960);
 
-			AssertThat((int)stripClock.time(), Equals(960));
-			AssertThat((int)videoClock.time(), Equals(960));
-			AssertThat((int)syncClock.time(), Equals(960));
+			AssertThat(stripClock.time(), Equals(960));
+			AssertThat(videoClock.time(), Equals(960));
+			AssertThat(syncClock.time(), Equals(960));
 
 			stripClock.setRate(1);
 			syncClock.setTime(1920);
 
-			AssertThat((int)stripClock.time(), Equals(960));
-			AssertThat((int)videoClock.time(), Equals(1920));
-			AssertThat((int)syncClock.time(), Equals(1920));
+			AssertThat(stripClock.time(), Equals(960));
+			AssertThat(videoClock.time(), Equals(1920));
+			AssertThat(syncClock.time(), Equals(1920));
 
 			syncClock.setTime(11520);
 
-			AssertThat((int)stripClock.time(), Equals(11520));
-			AssertThat((int)videoClock.time(), Equals(11520));
-			AssertThat((int)syncClock.time(), Equals(11520));
+			AssertThat(stripClock.time(), Equals(11520));
+			AssertThat(videoClock.time(), Equals(11520));
+			AssertThat(syncClock.time(), Equals(11520));
 		});
 
 		it("handle_sync_rae_change", []() {
